+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Log cancellations and host function activity [#320](https://github.com/hypermodeAI/runtime/pull/320)
 - Add namespaces to collections to isolate storage [#321](https://github.com/hypermodeAI/runtime/pull/321)
 - Use more direct approach for registering host functions [#322](https://github.com/hypermodeAI/runtime/pull/322)
+- Add getNamespaces host function for collections [#324]https://github.com/hypermodeAI/runtime/pull/324
 
 ## 2024-07-23 - Version 0.10.1